requirements.txt
torch
torchvision
ultralytics
gradio
pillow
numpy